requirements.txt
aiohttp==3.6.2
aiopg==1.0.0
async-timeout==3.0.1
attrs==20.1.0
chardet==3.0.4
idna==2.10
multidict==4.7.6
psycopg2-binary==2.8.5
SQLAlchemy==1.3.19
typing-extensions==3.7.4.3
yarl==1.5.1
